# api-schema/generated/openapi.yaml
openapi: 3.0.0
info:
  title: SubsQ API
  version: 0.0.0
...
paths:
...
  /api/v1/payment-methods:
    get:
      operationId: PaymentMethods_listPaymentMethods
      summary: List payment methods
      description: 支払い方法一覧を取得
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Models.PaymentMethodResponse'
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.UnauthorizedError'
      tags:
        - PaymentMethods
      security:
        - ApiKeyAuth: []
    post:
      operationId: PaymentMethods_createPaymentMethod
      summary: Create payment method
      description: 支払い方法を作成
      parameters: []
      responses:
        '201':
          description: The request has succeeded and a new resource has been created as a result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.PaymentMethodResponse'
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/Models.BadRequestError'
                  - $ref: '#/components/schemas/Models.UnauthorizedError'
      tags:
        - PaymentMethods
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Models.CreatePaymentMethodRequest'
      security:
        - ApiKeyAuth: []
    delete:
      operationId: PaymentMethods_deletePaymentMethods
      summary: Delete multiple payment methods
      description: 複数の支払い方法を削除
      parameters: []
      responses:
        '204':
          description: 'There is no content to send for this request, but the headers may be useful. '
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/Models.BadRequestError'
                  - $ref: '#/components/schemas/Models.ConflictError'
                  - $ref: '#/components/schemas/Models.UnauthorizedError'
      tags:
        - PaymentMethods
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Routes.DeletePaymentMethodsRequest'
      security:
        - ApiKeyAuth: []
  /api/v1/payment-methods/{id}:
    get:
      operationId: PaymentMethods_getPaymentMethod
      summary: Get payment method by ID
      description: 支払い方法詳細を取得
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/Models.uuid'
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.PaymentMethodResponse'
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/Models.NotFoundError'
                  - $ref: '#/components/schemas/Models.UnauthorizedError'
      tags:
        - PaymentMethods
      security:
        - ApiKeyAuth: []
    patch:
      operationId: PaymentMethods_updatePaymentMethod
      summary: Update payment method
      description: 支払い方法を更新
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/Models.uuid'
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.PaymentMethodResponse'
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/Models.BadRequestError'
                  - $ref: '#/components/schemas/Models.NotFoundError'
                  - $ref: '#/components/schemas/Models.UnauthorizedError'
      tags:
        - PaymentMethods
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Models.UpdatePaymentMethodRequest'
      security:
        - ApiKeyAuth: []
    delete:
      operationId: PaymentMethods_deletePaymentMethod
      summary: Delete payment method
      description: 支払い方法を削除
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/Models.uuid'
      responses:
        '204':
          description: 'There is no content to send for this request, but the headers may be useful. '
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/Models.ConflictError'
                  - $ref: '#/components/schemas/Models.NotFoundError'
                  - $ref: '#/components/schemas/Models.UnauthorizedError'
      tags:
        - PaymentMethods
      security:
        - ApiKeyAuth: []
  /api/v1/subscriptions:
    get:
      operationId: Subscriptions_listSubscriptions
      summary: List subscriptions
      description: サブスクリプション一覧を取得
      parameters: []
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.ListSubscriptionsResponse'
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.UnauthorizedError'
      tags:
        - Subscriptions
      security:
        - ApiKeyAuth: []
    post:
      operationId: Subscriptions_createSubscription
      summary: Create subscription
      description: サブスクリプションを作成
      parameters: []
      responses:
        '201':
          description: The request has succeeded and a new resource has been created as a result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.SubscriptionResponse'
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/Models.BadRequestError'
                  - $ref: '#/components/schemas/Models.UnauthorizedError'
      tags:
        - Subscriptions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Models.CreateSubscriptionRequest'
      security:
        - ApiKeyAuth: []
    delete:
      operationId: Subscriptions_deleteSubscriptions
      summary: Delete multiple subscriptions
      description: 複数のサブスクリプションを削除
      parameters: []
      responses:
        '204':
          description: 'There is no content to send for this request, but the headers may be useful. '
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/Models.BadRequestError'
                  - $ref: '#/components/schemas/Models.UnauthorizedError'
      tags:
        - Subscriptions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Routes.DeleteSubscriptionsRequest'
      security:
        - ApiKeyAuth: []
  /api/v1/subscriptions/{id}:
    get:
      operationId: Subscriptions_getSubscription
      summary: Get subscription by ID
      description: サブスクリプション詳細を取得
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/Models.uuid'
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.SubscriptionResponse'
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/Models.NotFoundError'
                  - $ref: '#/components/schemas/Models.UnauthorizedError'
      tags:
        - Subscriptions
      security:
        - ApiKeyAuth: []
    patch:
      operationId: Subscriptions_updateSubscription
      summary: Update subscription
      description: サブスクリプションを更新
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/Models.uuid'
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Models.SubscriptionResponse'
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/Models.BadRequestError'
                  - $ref: '#/components/schemas/Models.NotFoundError'
                  - $ref: '#/components/schemas/Models.UnauthorizedError'
      tags:
        - Subscriptions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Models.UpdateSubscriptionRequest'
      security:
        - ApiKeyAuth: []
    delete:
      operationId: Subscriptions_deleteSubscription
      summary: Delete subscription
      description: サブスクリプションを削除
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/Models.uuid'
      responses:
        '204':
          description: 'There is no content to send for this request, but the headers may be useful. '
        default:
          description: An unexpected error response.
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: '#/components/schemas/Models.NotFoundError'
                  - $ref: '#/components/schemas/Models.UnauthorizedError'
      tags:
        - Subscriptions
      security:
        - ApiKeyAuth: []
...
components:
  schemas:
    Models.BadRequestError:
      type: object
      required:
        - title
        - status
      properties:
        title:
          type: string
          enum:
            - Bad Request
        status:
          type: number
          enum:
            - 400
      allOf:
        - $ref: '#/components/schemas/Models.ErrorResponse'
      description: Bad Request エラー
...
    Models.ConflictError:
      type: object
      required:
        - title
        - status
      properties:
        title:
          type: string
          enum:
            - Conflict
        status:
          type: number
          enum:
            - 409
      allOf:
        - $ref: '#/components/schemas/Models.ErrorResponse'
      description: Conflict エラー
    Models.CreatePaymentMethodRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          description: 支払い方法名
      description: 支払い方法作成リクエスト
...
    Models.NotFoundError:
      type: object
      required:
        - title
        - status
      properties:
        title:
          type: string
          enum:
            - Not Found
        status:
          type: number
          enum:
            - 404
      allOf:
        - $ref: '#/components/schemas/Models.ErrorResponse'
      description: Not Found エラー
    Models.PaymentMethodResponse:
      type: object
      required:
        - id
        - userId
        - name
        - usageCount
        - createdAt
        - updatedAt
      properties:
        id:
          allOf:
            - $ref: '#/components/schemas/Models.uuid'
          description: 支払い方法ID
        userId:
          allOf:
            - $ref: '#/components/schemas/Models.uuid'
          description: ユーザーID
        name:
          type: string
          description: 支払い方法名
        usageCount:
          type: integer
          format: int32
          description: 使用中のサブスクリプション数
        createdAt:
          type: string
          format: date-time
          description: 作成日時
        updatedAt:
          type: string
          format: date-time
          description: 更新日時
      description: 支払い方法レスポンス
...
    Models.UnauthorizedError:
      type: object
      required:
        - title
        - status
      properties:
        title:
          type: string
          enum:
            - Unauthorized
        status:
          type: number
          enum:
            - 401
      allOf:
        - $ref: '#/components/schemas/Models.ErrorResponse'
      description: Unauthorized エラー
    Models.UpdatePaymentMethodRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          description: 支払い方法名
      description: 支払い方法更新リクエスト
...
    Models.uuid:
      type: string
      format: uuid
      description: UUID形式の文字列（再利用可能な共通スカラー）
    Routes.DeletePaymentMethodsRequest:
      type: object
      required:
        - ids
      properties:
        ids:
          type: array
          items:
            $ref: '#/components/schemas/Models.uuid'
    Routes.DeleteSubscriptionsRequest:
      type: object
      required:
        - ids
      properties:
        ids:
          type: array
          items:
            $ref: '#/components/schemas/Models.uuid'
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: cookie
      name: subsq_token
servers:
  - url: https://api.subsq-app.com
    description: Production server
    variables: {}
  - url: http://localhost:8080
    description: Development server
    variables: {}