# backend/internal/adapter/http/generated/openapi/server.gen.go
// Package openapi provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.6.0 DO NOT EDIT.
package openapi

import (
	"fmt"
	"net/http"
	"time"

	"github.com/labstack/echo/v4"
	"github.com/oapi-codegen/runtime"
	openapi_types "github.com/oapi-codegen/runtime/types"
)

const (
	ApiKeyAuthScopes = "ApiKeyAuth.Scopes"
)

// Defines values for ModelsBillingCycle.
const (
	ModelsBillingCycleMonthly ModelsBillingCycle = "monthly"
	ModelsBillingCycleYearly  ModelsBillingCycle = "yearly"
)

// Valid indicates whether the value is a known member of the ModelsBillingCycle enum.
func (e ModelsBillingCycle) Valid() bool {
	switch e {
	case ModelsBillingCycleMonthly:
		return true
	case ModelsBillingCycleYearly:
		return true
	default:
		return false
	}
}

// ModelsBadRequestError RFC 7807 Problem Details エラーレスポンス
type ModelsBadRequestError = ModelsErrorResponse

// ModelsBillingCycle 請求サイクル
type ModelsBillingCycle string

// ModelsConflictError RFC 7807 Problem Details エラーレスポンス
type ModelsConflictError = ModelsErrorResponse

// ModelsCreatePaymentMethodRequest 支払い方法作成リクエスト
type ModelsCreatePaymentMethodRequest struct {
	// Name 支払い方法名
	Name string `json:"name"`
}

// ModelsCreateSubscriptionRequest サブスクリプション作成リクエスト
type ModelsCreateSubscriptionRequest struct {
	// Amount 金額（円）
	Amount int64 `json:"amount"`

	// BaseDate サブスクリプション開始日（ISO 8601 UTC）
	BaseDate time.Time `json:"baseDate"`

	// BillingCycle 請求サイクル
	BillingCycle ModelsBillingCycle `json:"billingCycle"`

	// Memo メモ
	Memo *string `json:"memo,omitempty"`

	// PaymentMethodId 支払い方法ID
	PaymentMethodId *ModelsUuid `json:"paymentMethodId,omitempty"`

	// ServiceName サービス名
	ServiceName string `json:"serviceName"`
}

// ModelsErrorResponse RFC 7807 Problem Details エラーレスポンス
type ModelsErrorResponse struct {
	// Detail エラーの詳細説明（RFC 7807: SHOULD include、必須ではない）
	Detail *string `json:"detail,omitempty"`

	// Instance エラーが発生したリソースの URI（オプション）
	Instance *string `json:"instance,omitempty"`

	// Status HTTP ステータスコード
	Status int32 `json:"status"`

	// Title エラーの短い説明
	Title string `json:"title"`

	// Type エラータイプの URI（RFC 7807: URI参照）
	Type string `json:"type"`
}

// ModelsForbiddenError RFC 7807 Problem Details エラーレスポンス
type ModelsForbiddenError = ModelsErrorResponse

// ModelsListSubscriptionsResponse サブスクリプション一覧レスポンス
type ModelsListSubscriptionsResponse struct {
	// Subscriptions サブスクリプション一覧
	Subscriptions []ModelsSubscriptionResponse `json:"subscriptions"`

	// Summary 合計情報
	Summary ModelsSubscriptionListSummary `json:"summary"`
}

// ModelsLogoutResponse ログアウトレスポンス
type ModelsLogoutResponse struct {
	Message string `json:"message"`
}

// ModelsNotFoundError RFC 7807 Problem Details エラーレスポンス
type ModelsNotFoundError = ModelsErrorResponse

// ModelsPaymentMethodResponse 支払い方法レスポンス
type ModelsPaymentMethodResponse struct {
	// CreatedAt 作成日時
	CreatedAt time.Time `json:"createdAt"`

	// Id 支払い方法ID
	Id ModelsUuid `json:"id"`

	// Name 支払い方法名
	Name string `json:"name"`

	// UpdatedAt 更新日時
	UpdatedAt time.Time `json:"updatedAt"`

	// UsageCount 使用中のサブスクリプション数
	UsageCount int32 `json:"usageCount"`

	// UserId ユーザーID
	UserId ModelsUuid `json:"userId"`
}

// ModelsPaymentMethodSummary 支払い方法の簡易情報（サブスクに埋め込まれる）
type ModelsPaymentMethodSummary struct {
	// Id 支払い方法ID
	Id ModelsUuid `json:"id"`

	// Name 支払い方法名
	Name string `json:"name"`
}

// ModelsSubscriptionListSummary サブスクリプション一覧レスポンスのサマリー
type ModelsSubscriptionListSummary struct {
	// Count 件数
	Count int32 `json:"count"`

	// MonthlyTotal 月額合計（円）
	MonthlyTotal int64 `json:"monthlyTotal"`

	// YearlyTotal 年額合計（円）
	YearlyTotal int64 `json:"yearlyTotal"`
}

// ModelsSubscriptionResponse サブスクリプションレスポンス
type ModelsSubscriptionResponse struct {
	// Amount 金額（円）
	Amount int64 `json:"amount"`

	// BaseDate サブスクリプション開始日（ISO 8601 UTC）
	BaseDate time.Time `json:"baseDate"`

	// BillingCycle 請求サイクル
	BillingCycle ModelsBillingCycle `json:"billingCycle"`

	// CreatedAt 作成日時
	CreatedAt time.Time `json:"createdAt"`

	// Id サブスクリプションID
	Id ModelsUuid `json:"id"`

	// Memo メモ
	Memo *string `json:"memo,omitempty"`

	// MonthlyAmount 月額換算金額（円）
	MonthlyAmount int64 `json:"monthlyAmount"`

	// NextBillingDate 次回請求日（YYYY-MM-DD、時刻を持たない日付のみの値）
	NextBillingDate openapi_types.Date `json:"nextBillingDate"`

	// PaymentMethod 支払い方法情報
	PaymentMethod *ModelsPaymentMethodSummary `json:"paymentMethod,omitempty"`

	// PaymentMethodId 支払い方法ID
	PaymentMethodId *ModelsUuid `json:"paymentMethodId,omitempty"`

	// ServiceName サービス名
	ServiceName string `json:"serviceName"`

	// UpdatedAt 更新日時
	UpdatedAt time.Time `json:"updatedAt"`

	// UserId ユーザーID
	UserId ModelsUuid `json:"userId"`

	// YearlyAmount 年額換算金額（円）
	YearlyAmount int64 `json:"yearlyAmount"`
}

// ModelsUnauthorizedError RFC 7807 Problem Details エラーレスポンス
type ModelsUnauthorizedError = ModelsErrorResponse

// ModelsUpdatePaymentMethodRequest 支払い方法更新リクエスト
type ModelsUpdatePaymentMethodRequest struct {
	// Name 支払い方法名
	Name *string `json:"name,omitempty"`
}

// ModelsUpdateSubscriptionRequest サブスクリプション更新リクエスト
type ModelsUpdateSubscriptionRequest struct {
	// Amount 金額（円）
	Amount *int64 `json:"amount,omitempty"`

	// BaseDate サブスクリプション開始日（ISO 8601 UTC）
	BaseDate *time.Time `json:"baseDate,omitempty"`

	// BillingCycle 請求サイクル
	BillingCycle *ModelsBillingCycle `json:"billingCycle,omitempty"`

	// Memo メモ
	Memo *string `json:"memo,omitempty"`

	// PaymentMethodId 支払い方法ID
	PaymentMethodId *ModelsUuid `json:"paymentMethodId,omitempty"`

	// ServiceName サービス名
	ServiceName *string `json:"serviceName,omitempty"`
}

// ModelsUpdateUserRequest ユーザープロフィール更新リクエスト
type ModelsUpdateUserRequest struct {
	// Name 表示名
	Name *string `json:"name,omitempty"`

	// Thumbnail プロフィール画像URL（null で削除）
	Thumbnail *string `json:"thumbnail,omitempty"`
}

// ModelsUserResponse ユーザー情報レスポンス
type ModelsUserResponse struct {
	// CreatedAt 作成日時
	CreatedAt time.Time `json:"createdAt"`

	// Email メールアドレス
	Email openapi_types.Email `json:"email"`

	// Id ユーザーID
	Id ModelsUuid `json:"id"`

	// Name 表示名
	Name string `json:"name"`

	// Provider OAuthプロバイダー（例: google）
	Provider string `json:"provider"`

	// ProviderAccountId プロバイダー側のアカウントID
	ProviderAccountId string `json:"providerAccountId"`

	// Thumbnail プロフィール画像URL
	Thumbnail *string `json:"thumbnail,omitempty"`

	// UpdatedAt 更新日時
	UpdatedAt time.Time `json:"updatedAt"`
}

// ModelsUuid UUID形式の文字列（再利用可能な共通スカラー）
type ModelsUuid = openapi_types.UUID

// RoutesDeletePaymentMethodsRequest defines model for Routes.DeletePaymentMethodsRequest.
type RoutesDeletePaymentMethodsRequest struct {
	Ids []ModelsUuid `json:"ids"`
}

// RoutesDeleteSubscriptionsRequest defines model for Routes.DeleteSubscriptionsRequest.
type RoutesDeleteSubscriptionsRequest struct {
	Ids []ModelsUuid `json:"ids"`
}

// PaymentMethodsDeletePaymentMethodsJSONRequestBody defines body for PaymentMethodsDeletePaymentMethods for application/json ContentType.
type PaymentMethodsDeletePaymentMethodsJSONRequestBody = RoutesDeletePaymentMethodsRequest

// PaymentMethodsCreatePaymentMethodJSONRequestBody defines body for PaymentMethodsCreatePaymentMethod for application/json ContentType.
type PaymentMethodsCreatePaymentMethodJSONRequestBody = ModelsCreatePaymentMethodRequest

// PaymentMethodsUpdatePaymentMethodJSONRequestBody defines body for PaymentMethodsUpdatePaymentMethod for application/json ContentType.
type PaymentMethodsUpdatePaymentMethodJSONRequestBody = ModelsUpdatePaymentMethodRequest

// SubscriptionsDeleteSubscriptionsJSONRequestBody defines body for SubscriptionsDeleteSubscriptions for application/json ContentType.
type SubscriptionsDeleteSubscriptionsJSONRequestBody = RoutesDeleteSubscriptionsRequest

// SubscriptionsCreateSubscriptionJSONRequestBody defines body for SubscriptionsCreateSubscription for application/json ContentType.
type SubscriptionsCreateSubscriptionJSONRequestBody = ModelsCreateSubscriptionRequest

// SubscriptionsUpdateSubscriptionJSONRequestBody defines body for SubscriptionsUpdateSubscription for application/json ContentType.
type SubscriptionsUpdateSubscriptionJSONRequestBody = ModelsUpdateSubscriptionRequest

// UsersUpdateCurrentUserJSONRequestBody defines body for UsersUpdateCurrentUser for application/json ContentType.
type UsersUpdateCurrentUserJSONRequestBody = ModelsUpdateUserRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Logout
	// (POST /api/v1/auth/logout)
	AuthLogout(ctx echo.Context) error
	// Delete multiple payment methods
	// (DELETE /api/v1/payment-methods)
	PaymentMethodsDeletePaymentMethods(ctx echo.Context) error
	// List payment methods
	// (GET /api/v1/payment-methods)
	PaymentMethodsListPaymentMethods(ctx echo.Context) error
	// Create payment method
	// (POST /api/v1/payment-methods)
	PaymentMethodsCreatePaymentMethod(ctx echo.Context) error
	// Delete payment method
	// (DELETE /api/v1/payment-methods/{id})
	PaymentMethodsDeletePaymentMethod(ctx echo.Context, id ModelsUuid) error
	// Get payment method by ID
	// (GET /api/v1/payment-methods/{id})
	PaymentMethodsGetPaymentMethod(ctx echo.Context, id ModelsUuid) error
	// Update payment method
	// (PATCH /api/v1/payment-methods/{id})
	PaymentMethodsUpdatePaymentMethod(ctx echo.Context, id ModelsUuid) error
	// Delete multiple subscriptions
	// (DELETE /api/v1/subscriptions)
	SubscriptionsDeleteSubscriptions(ctx echo.Context) error
	// List subscriptions
	// (GET /api/v1/subscriptions)
	SubscriptionsListSubscriptions(ctx echo.Context) error
	// Create subscription
	// (POST /api/v1/subscriptions)
	SubscriptionsCreateSubscription(ctx echo.Context) error
	// Delete subscription
	// (DELETE /api/v1/subscriptions/{id})
	SubscriptionsDeleteSubscription(ctx echo.Context, id ModelsUuid) error
	// Get subscription by ID
	// (GET /api/v1/subscriptions/{id})
	SubscriptionsGetSubscription(ctx echo.Context, id ModelsUuid) error
	// Update subscription
	// (PATCH /api/v1/subscriptions/{id})
	SubscriptionsUpdateSubscription(ctx echo.Context, id ModelsUuid) error
	// Delete current user
	// (DELETE /api/v1/users/me)
	UsersDeleteCurrentUser(ctx echo.Context) error
	// Get current user
	// (GET /api/v1/users/me)
	UsersGetCurrentUser(ctx echo.Context) error
	// Update current user
	// (PATCH /api/v1/users/me)
	UsersUpdateCurrentUser(ctx echo.Context) error
}

// ServerInterfaceWrapper converts echo contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler ServerInterface
}

// AuthLogout converts echo context to params.
func (w *ServerInterfaceWrapper) AuthLogout(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.AuthLogout(ctx)
	return err
}

// PaymentMethodsDeletePaymentMethods converts echo context to params.
func (w *ServerInterfaceWrapper) PaymentMethodsDeletePaymentMethods(ctx echo.Context) error {
	var err error

	ctx.Set(ApiKeyAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.PaymentMethodsDeletePaymentMethods(ctx)
	return err
}

// PaymentMethodsListPaymentMethods converts echo context to params.
func (w *ServerInterfaceWrapper) PaymentMethodsListPaymentMethods(ctx echo.Context) error {
	var err error

	ctx.Set(ApiKeyAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.PaymentMethodsListPaymentMethods(ctx)
	return err
}

// PaymentMethodsCreatePaymentMethod converts echo context to params.
func (w *ServerInterfaceWrapper) PaymentMethodsCreatePaymentMethod(ctx echo.Context) error {
	var err error

	ctx.Set(ApiKeyAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.PaymentMethodsCreatePaymentMethod(ctx)
	return err
}

// PaymentMethodsDeletePaymentMethod converts echo context to params.
func (w *ServerInterfaceWrapper) PaymentMethodsDeletePaymentMethod(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "id" -------------
	var id ModelsUuid

	err = runtime.BindStyledParameterWithOptions("simple", "id", ctx.Param("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "uuid"})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter id: %s", err))
	}

	ctx.Set(ApiKeyAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.PaymentMethodsDeletePaymentMethod(ctx, id)
	return err
}

// PaymentMethodsGetPaymentMethod converts echo context to params.
func (w *ServerInterfaceWrapper) PaymentMethodsGetPaymentMethod(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "id" -------------
	var id ModelsUuid

	err = runtime.BindStyledParameterWithOptions("simple", "id", ctx.Param("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "uuid"})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter id: %s", err))
	}

	ctx.Set(ApiKeyAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.PaymentMethodsGetPaymentMethod(ctx, id)
	return err
}

// PaymentMethodsUpdatePaymentMethod converts echo context to params.
func (w *ServerInterfaceWrapper) PaymentMethodsUpdatePaymentMethod(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "id" -------------
	var id ModelsUuid

	err = runtime.BindStyledParameterWithOptions("simple", "id", ctx.Param("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "uuid"})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter id: %s", err))
	}

	ctx.Set(ApiKeyAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.PaymentMethodsUpdatePaymentMethod(ctx, id)
	return err
}

// SubscriptionsDeleteSubscriptions converts echo context to params.
func (w *ServerInterfaceWrapper) SubscriptionsDeleteSubscriptions(ctx echo.Context) error {
	var err error

	ctx.Set(ApiKeyAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.SubscriptionsDeleteSubscriptions(ctx)
	return err
}

// SubscriptionsListSubscriptions converts echo context to params.
func (w *ServerInterfaceWrapper) SubscriptionsListSubscriptions(ctx echo.Context) error {
	var err error

	ctx.Set(ApiKeyAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.SubscriptionsListSubscriptions(ctx)
	return err
}

// SubscriptionsCreateSubscription converts echo context to params.
func (w *ServerInterfaceWrapper) SubscriptionsCreateSubscription(ctx echo.Context) error {
	var err error

	ctx.Set(ApiKeyAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.SubscriptionsCreateSubscription(ctx)
	return err
}

// SubscriptionsDeleteSubscription converts echo context to params.
func (w *ServerInterfaceWrapper) SubscriptionsDeleteSubscription(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "id" -------------
	var id ModelsUuid

	err = runtime.BindStyledParameterWithOptions("simple", "id", ctx.Param("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "uuid"})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter id: %s", err))
	}

	ctx.Set(ApiKeyAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.SubscriptionsDeleteSubscription(ctx, id)
	return err
}

// SubscriptionsGetSubscription converts echo context to params.
func (w *ServerInterfaceWrapper) SubscriptionsGetSubscription(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "id" -------------
	var id ModelsUuid

	err = runtime.BindStyledParameterWithOptions("simple", "id", ctx.Param("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "uuid"})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter id: %s", err))
	}

	ctx.Set(ApiKeyAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.SubscriptionsGetSubscription(ctx, id)
	return err
}

// SubscriptionsUpdateSubscription converts echo context to params.
func (w *ServerInterfaceWrapper) SubscriptionsUpdateSubscription(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "id" -------------
	var id ModelsUuid

	err = runtime.BindStyledParameterWithOptions("simple", "id", ctx.Param("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "uuid"})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter id: %s", err))
	}

	ctx.Set(ApiKeyAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.SubscriptionsUpdateSubscription(ctx, id)
	return err
}

// UsersDeleteCurrentUser converts echo context to params.
func (w *ServerInterfaceWrapper) UsersDeleteCurrentUser(ctx echo.Context) error {
	var err error

	ctx.Set(ApiKeyAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.UsersDeleteCurrentUser(ctx)
	return err
}

// UsersGetCurrentUser converts echo context to params.
func (w *ServerInterfaceWrapper) UsersGetCurrentUser(ctx echo.Context) error {
	var err error

	ctx.Set(ApiKeyAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.UsersGetCurrentUser(ctx)
	return err
}

// UsersUpdateCurrentUser converts echo context to params.
func (w *ServerInterfaceWrapper) UsersUpdateCurrentUser(ctx echo.Context) error {
	var err error

	ctx.Set(ApiKeyAuthScopes, []string{})

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.UsersUpdateCurrentUser(ctx)
	return err
}

// This is a simple interface which specifies echo.Route addition functions which
// are present on both echo.Echo and echo.Group, since we want to allow using
// either of them for path registration
type EchoRouter interface {
	CONNECT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	DELETE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	GET(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	HEAD(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	OPTIONS(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PATCH(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	POST(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PUT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	TRACE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
}

// RegisterHandlers adds each server route to the EchoRouter.
func RegisterHandlers(router EchoRouter, si ServerInterface) {
	RegisterHandlersWithBaseURL(router, si, "")
}

// Registers handlers, and prepends BaseURL to the paths, so that the paths
// can be served under a prefix.
func RegisterHandlersWithBaseURL(router EchoRouter, si ServerInterface, baseURL string) {

	wrapper := ServerInterfaceWrapper{
		Handler: si,
	}

	router.POST(baseURL+"/api/v1/auth/logout", wrapper.AuthLogout)
	router.DELETE(baseURL+"/api/v1/payment-methods", wrapper.PaymentMethodsDeletePaymentMethods)
	router.GET(baseURL+"/api/v1/payment-methods", wrapper.PaymentMethodsListPaymentMethods)
	router.POST(baseURL+"/api/v1/payment-methods", wrapper.PaymentMethodsCreatePaymentMethod)
	router.DELETE(baseURL+"/api/v1/payment-methods/:id", wrapper.PaymentMethodsDeletePaymentMethod)
	router.GET(baseURL+"/api/v1/payment-methods/:id", wrapper.PaymentMethodsGetPaymentMethod)
	router.PATCH(baseURL+"/api/v1/payment-methods/:id", wrapper.PaymentMethodsUpdatePaymentMethod)
	router.DELETE(baseURL+"/api/v1/subscriptions", wrapper.SubscriptionsDeleteSubscriptions)
	router.GET(baseURL+"/api/v1/subscriptions", wrapper.SubscriptionsListSubscriptions)
	router.POST(baseURL+"/api/v1/subscriptions", wrapper.SubscriptionsCreateSubscription)
	router.DELETE(baseURL+"/api/v1/subscriptions/:id", wrapper.SubscriptionsDeleteSubscription)
	router.GET(baseURL+"/api/v1/subscriptions/:id", wrapper.SubscriptionsGetSubscription)
	router.PATCH(baseURL+"/api/v1/subscriptions/:id", wrapper.SubscriptionsUpdateSubscription)
	router.DELETE(baseURL+"/api/v1/users/me", wrapper.UsersDeleteCurrentUser)
	router.GET(baseURL+"/api/v1/users/me", wrapper.UsersGetCurrentUser)
	router.PATCH(baseURL+"/api/v1/users/me", wrapper.UsersUpdateCurrentUser)

}